% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addExtent.R
\name{addExtent}
\alias{addExtent}
\title{view spatial objects interactively}
\usage{
addExtent(x, map = NULL, map.types = c("OpenStreetMap",
  "Esri.WorldImagery"), verbose = FALSE, ...)
}
\description{
view spatial objects interactively
}

