% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addMapLayer.R
\name{addMapLayer}
\alias{addMapLayer}
\title{add a layer to a leaflet map using mapView}
\usage{
addMapLayer(lay, map)
}
\arguments{
\item{map}{the map to which the layer should be added}

\item{object}{an object that can be passed to \code{\link{mapView}}}

\item{...}{additional arguments passed on to \code{\link{mapView}}}
}
\description{
this function adds a layer to an interactive GIS-like view produced
with \code{\link{mapView}}
}
\examples{
### raster data ###
library(sp)
library(raster)

data(meuse.grid)
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) = TRUE
meuse_rst <- stack(meuse.grid)

m1 <- mapView(meuse_rst[[3]])
m1

### point vector data ###
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

# all layers of meuse
addMapLayer(meuse, m1)
}
\author{
Tim Appelhans
}

