% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStarsRGB.R
\name{addStarsRGB}
\alias{addStarsRGB}
\title{Add a RGB stars image as a layer}
\usage{
addStarsRGB(map, x, r = 3, g = 2, b = 1, quantiles = c(0.02, 0.98),
  maxpixels = mapviewGetOption("mapview.maxpixels"),
  na.color = mapviewGetOption("na.color"), method = c("bilinear",
  "ngb"), ...)
}
\arguments{
\item{map}{a map widget object created from `leaflet()``}

\item{x}{a RasterBrick or RasterStack}

\item{r}{integer. Index of the Red channel/band, between 1 and nlayers(x)}

\item{g}{integer. Index of the Green channel/band, between 1 and nlayers(x)}

\item{b}{integer. Index of the Blue channel/band, between 1 and nlayers(x)}

\item{quantiles}{the upper and lower quantiles used for color stretching. If set to NULL, no stretching is applied.}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{na.color}{the color to be used for NA pixels}

\item{method}{Method used to compute
values for the resampled layer that is passed on to leaflet. mapview does
projection on-the-fly to ensure correct display and therefore needs to know
how to do this projection. The default is 'bilinear' (bilinear interpolation),
which is appropriate for continuous variables. The other option, 'ngb'
(nearest neighbor), is useful for categorical variables.}

\item{...}{additional arguments passed on to \code{\link{addStarsImage}}}
}
\description{
Create a Red-Green-Blue image overlay from a \code{stars} object
based on three layers.
This is an adaptation of \code{mapview::viewRGB} function written in order to
give the possibility to add a RGB image to a leaflet / leafletProxy object.
}
\examples{
if (interactive()) {
  library(raster)
  library(stars)
  library(plainview)
  library(leflet)

  poppendorf <- st_as_stars(plainview::poppendorf)
  leaflet() \%>\%
    addTiles(group = "OpenStreetMap") \%>\%
    addStarsRGB(poppendorf, 4,3,2, group = "True colours") \%>\%
    addStarsRGB(poppendorf, 5,4,3, group = "False colours") \%>\%
    addLayersControl(
      baseGroups = c("Satellite"),
      overlayGroups = c("True colours", "False colours"),
    )
}

}
\author{
Tim Appelhans, Luigi Ranghetti
}
