% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{addTiledRasterImage}
\alias{addTiledRasterImage}
\title{Add a tiled image}
\usage{
addTiledRasterImage(map, x, minzoom = 0, maxzoom,
  color = viridisLite::inferno, at, na.color = "#BEBEBE")
}
\description{

}
