% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{atlStorms2005}
\alias{atlStorms2005}
\title{Atlantic Ocean storms 2005}
\format{\code{sp::SpatialLinesDataFrame}}
\description{
Atlantic Ocean storms 2005
}
\details{


This dataset contains storm tracks for selected storms
in the Atlantic Ocean basin for the year 2005
}

