% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brewControls.R
\name{brewControls}
\alias{brewControls}
\alias{brewPopupRow}
\alias{brewPopupRowAlt}
\alias{brewPopupTable}
\title{Convenience functions for brew plugin}
\usage{
brewPopupTable(x)

brewPopupRow(col.name, value)

brewPopupRowAlt(col.name, value)
}
\arguments{
\item{col.name}{the name of the column of the attribute table}

\item{value}{the corresponding value from the attribute table}
}
\description{
Convenience functions for brew plugin
}
\section{Functions}{
\itemize{
\item \code{brewPopupTable}: create popup table for sp objects

\item \code{brewPopupRow}: create popup table odd row for sp objects

\item \code{brewPopupRowAlt}: create popup table even row for sp objects
}}
\author{
Tim Appelhans, Florian Detsch
}

