% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvecView.R
\name{bViewOutput}
\alias{bViewOutput}
\title{Widget output function for use in Shiny}
\usage{
bViewOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

