% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{campsQ2}
\alias{campsQ2}
\title{Antartic Facilities}
\format{\code{sp::SpatialPointsDataFrame}}
\description{
Antartic facilities. Type of facility attribute is ""feature_ty"
}
\details{
The data show the antartic facilities (2012) as provided by \href{https://www.comnap.aq}{comnap} via the \href{www.quantartica.org}{Quantartica2} data compilation that is compiled by the \href{http://www.npolar.no/en}{Norwegian Polar Institute}
}

