% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeView_RasterStack}
\alias{cubeView_RasterStack}
\title{View a RasterStack as 3-dimensional cube.}
\usage{
cubeView_RasterStack(r)
}
\description{
see details of cubeView: \link{cubeViewRaw}
}
\author{
Stephan Woellauer
}

