% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeViewOutput}
\alias{cubeViewOutput}
\title{Widget output function for use in Shiny}
\usage{
cubeViewOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

