% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeViewRaw}
\alias{cubeViewRaw}
\title{View a cube of 3-dimensional data filled with points (voxels).}
\usage{
cubeViewRaw(grey = NULL, red = NULL, green = NULL, blue = NULL, x_size,
  y_size, z_size, width = NULL, height = NULL)
}
\arguments{
\item{grey}{optional integer vector with 0 <= value <= 255.}

\item{red}{optional integer vector with 0 <= value <= 255.}

\item{green}{optional integer vector with 0 <= value <= 255.}

\item{blue}{optional integer vector with 0 <= value <= 255.}

\item{x_size}{integer. size of x-dimension}

\item{y_size}{integer. size of y-dimension}

\item{z_size}{integer. size of z-dimension}
}
\description{
A variation of Hovmoeller diagram: Each voxel is colored with a RGB-color (or grey) value.
}
\details{
The cube faces show a selectable layer of data within the cube.

The visible layers are alterable by keys:

x-axis: LEFT / RIGHT arrow key

y-axis: DOWN / UP arrow key

z-axis: PAGE_DOWN / PAGE_UP key

Note: Because of key focus issues key-press-events are not always recognised within RStudio at Windows.
In this case open the view in a web-browser (RStudio button: "show in new window").


Press and hold left mouse-button to rotate the cube.

Press and hold right mouse-button to move the cube.

Spin mouse-wheel or press and hold middle mouse-button and move mouse down/up to zoom the cube.

The color resp. grey vectors contain sequentially values of each voxel. So each vector is length == x_size * y_size * z_size.
Color component values overwrite grey values.

Sequence of coordinates (x,y,z) for values in vectors:

(1,1,1), (2,1,1), (3,1,1), ... (1,2,1), (2,2,1), (3,2,1), ... (1,1,2), (2,1,2), (3,1,2), ...
}
\author{
Stephan Woellauer
}

