% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{gadmCHE}
\alias{gadmCHE}
\title{Administrative borders of Switzerland (level 2)}
\format{\code{sp::SpatialPolygonsDataFrame}}
\source{
\url{http://gadm.org}
}
\description{
Administrative borders of Switzerland (level 2)
}
\details{


This dataset comes from \url{http://gadm.org}.
It was downloaded using \code{\link{getData}}.
}

