% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapControls.R
\name{getLayerControlEntriesFromMap}
\alias{getLayerControlEntriesFromMap}
\title{convenience functions for working with leaflet maps}
\usage{
getLayerControlEntriesFromMap(m)
}
\arguments{
\item{m}{a leaflet map}
}
\description{
convenience functions for working with leaflet maps
}
\author{
Tim Appelhans
}

