% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{hillshadeCHE}
\alias{hillshadeCHE}
\title{Hillshade raster of Switzerland}
\format{\code{"RasterLayer-class"}}
\description{
Hillshade raster of Switzerland
}
\details{
This dataset was created following the example
at \code{\link{hillShade}}.
}

