% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{kiliLS8}
\alias{kiliLS8}
\title{Landsat 8 detail of Mt Kilimanjaro}
\format{\code{"RasterBrick-class"} with 11 bands.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\description{
Landsat 8 detail of Mt Kilimanjaro showing an elevational
cross-section from savannah to alpine helichrysum vegetation
}
\details{


Use of this data requires your agreement to the USGS regulations on
using Landsat data.
}

