% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{kiliNDVI}
\alias{kiliNDVI}
\title{16-day NDVI of Kilimanjaro}
\format{\code{"RasterStack-class"}}
\description{
16-day NDVI of Kilimanjaro from January to December 2013
}
\details{
16-day Aqua-MODIS NDVI (MYD13Q1, Collection 6) from 2013,
quality-controlled and Whittaker-smoothed as described in Detsch et al.
(2016; see References).
}
\references{
Detsch F, Otte I, Appelhans T, Hemp A, Nauss T (2016) Seasonal
and long-term vegetation dynamics from 1-km GIMMS-based NDVI time series at
Mt. Kilimanjaro, Tanzania. Remote Sensing of Environment 178: 70--83,
doi:10.1016/j.rse.2016.03.007
}

