% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{knit_print.mapview}
\alias{knit_print.mapview}
\title{Print functions for mapview objects used in knitr}
\usage{
knit_print.mapview(x, ...)
}
\arguments{
\item{x}{A mapview object}

\item{...}{further arguments passed on to \code{\link[knitr]{knit_print}}}
}
\description{
Print functions for mapview objects used in knitr
}

