% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticeView.R
\name{latticeView}
\alias{latticeView}
\alias{latticeview}
\title{View two or more possibly synchronised mapview or leaflet maps}
\usage{
latticeView(..., ncol = 2, sync = "all", sync.curser = TRUE,
  no.initial.sync = TRUE)

latticeview(...)
}
\arguments{
\item{...}{any number of mapview or leaflet objects or a list thereof}

\item{ncol}{how many columns should be plotted}

\item{sync}{whether to synchronise zoom and pan for certain elements.
Possible values are "all" (default) to sync all maps, "none" to disable
synchronisation or a list of panel numbers, e.g. \code{list(c(1, 3), c(2, 4))}
will synchronise panels 1 & 3 and panels 2 & 4. Panels are drawn from top right
to bottom left.}

\item{sync.curser}{whether to show curser position in synced panels (default TRUE).}

\item{no.initial.sync}{whether to sync the initial view (default TRUE).}
}
\description{
This function produces a view of two or more maps. It is possible to sync
any combination of panels or all or none.
}
\section{Functions}{
\itemize{
\item \code{latticeview}: alias for ease of typing
}}
\examples{
\dontrun{
library(sp)

data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

## view different aspects of same data set
m1 <- mapview(meuse, zcol = "soil", burst = TRUE)
m2 <- mapview(meuse, zcol = "lead")
m3 <- mapview(meuse, zcol = "landuse")
m4 <- mapview(meuse, zcol = "dist.m")

latticeView(m1, m2, m3, m4) # 4 panels
latticeView(m1, m2, m3, m4, sync.curser = FALSE) # 4 panels
latticeView(m1, m2) # 2 panels, split vertical
latticeView(m1, m2, ncol = 1) # 2 panels split horizontal
latticeView(m1, m2, m3, m4, sync = list(c(1, 2), c(3, 4))) # individual syncing
latticeView(m1, m2, m3, m4, sync = list(c(1, 2, 4)))

## view all layers of raster stack - not very resposive!!
map_list <- lapply(seq(nlayers(poppendorf)), function(i) {
  mapview(poppendorf[[i]], layer.name = names(poppendorf)[i])
})

latticeView(map_list, ncol = 3)

## view multiple data sets
m1 <- mapview(meuse, zcol = "soil", burst = TRUE)
m2 <- mapview(atlStorms2005, zcol = "Name")
m3 <- mapview(breweries91)
m4 <- mapview(gadmCHE, color = "cornflowerblue")

latticeView(m1, m2, m3, m4, sync = "none") # not synced
latticeView(m1, m2, m3, m4) # synced
latticeView(m1, m2, m3, m4, no.initial.sync = FALSE) # all maps zoomed to m4 extent

}

}

