% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leafletControls.R
\name{leafletControls}
\alias{appendMapCallEntries}
\alias{checkAdjustProjection}
\alias{circleRadius}
\alias{getLayerControlEntriesFromMap}
\alias{getLayerNamesFromMap}
\alias{getMapCalls}
\alias{getProviderTileEntriesFromMap}
\alias{getProviderTileNamesFromMap}
\alias{initBaseMaps}
\alias{initMap}
\alias{layerName}
\alias{layers2bHidden}
\alias{leafletControls}
\alias{mapViewLayersControl}
\alias{rasterCheckAdjustProjection}
\alias{removeDuplicatedMapCalls}
\alias{scaleCoordinates}
\alias{scaleExtent}
\alias{scaleLinesCoordinates}
\alias{scalePolygonsCoordinates}
\alias{spCheckAdjustProjection}
\alias{spCheckObject}
\alias{updateLayerControlNames}
\title{Convenience functions for working with spatial objects and leaflet maps}
\usage{
getLayerControlEntriesFromMap(map)

getLayerNamesFromMap(map)

getProviderTileEntriesFromMap(map)

getProviderTileNamesFromMap(map)

updateLayerControlNames(map1, map2)

layers2bHidden(map)

getMapCalls(map)

appendMapCallEntries(map1, map2)

removeDuplicatedMapCalls(map)

rasterCheckAdjustProjection(x, maxpixels)

initBaseMaps(map.types)

initMap(map, map.types, proj4str)

scaleCoordinates(x.coords, y.coords)

scaleExtent(x)

scalePolygonsCoordinates(x)

scaleLinesCoordinates(x)

spCheckAdjustProjection(x)

checkAdjustProjection(x, maxpixels)

mapViewLayersControl(map, map.types, names)

layerName()

circleRadius(x, radius)

spCheckObject(x, verbose)
}
\arguments{
\item{map}{a leaflet map}

\item{map1}{map to be modified}

\item{map2}{map to get modification from}

\item{x}{a Raster* or Spatial* object}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{map.types}{map types to be used as backgraound maps}

\item{proj4str}{\code{\link{proj4string}} of the spatial objects to be viewed}

\item{x.coords}{vector of x coordinates}

\item{y.coords}{vector of y coordinates}

\item{names}{names of the layer groups to be added to control button}

\item{radius}{either an integer specifying circle radius or the name
(character) of one of the columns in the SpatialPointsDataFrame
(attribute table)}

\item{verbose}{whether details should be printed to the console}
}
\description{
Convenience functions for working with spatial objects and leaflet maps
}
\author{
Tim Appelhans
}

