% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localTiles3031.R
\name{localTiles3031}
\alias{localTiles3031}
\title{create epsg3031 projected local tiles for leaflet and maps it}
\usage{
servr::httd("~/proj/makeTile/etopo/tiles",daemon=TRUE)
localTiles3031(x="http://localhost:4321/{z}/{x}/{y}.png")
}
\arguments{
\item{x}{list of local/remote urls pointing to the tiles to serve}

\item{zoom}{number of tile levels dont count the zero level}

\item{xname}{list of layer names}

\item{tileSize}{sie of the tiles in pixel default is 256,}

\item{attribution}{of the layer}
}
\description{
localTiles3031 creates and maps a tiles from a given raster or
  GDAL object. Up to know there is no way to initialize a htmlwidget object
  dynamically. So we have to create static versions for the region on higher
  latitude than +/- 85 degree
}
\examples{
\dontrun{
 ### due to a lot of incompatibiliets it is necesary to serve the local tiles aat local host
 ### we just use the servr package to do so.
  require(servr)
  # we serve the directory that we did use with see \\code{\\link{makeTile}}

  servr::httd("~/proj/makeTile/etopo/tiles",daemon=TRUE)
  # and we map it
  localTiles3031(x=c("http://localhost:4321/{z}/{x}/{y}.png","http://localhost:4321/{z}/{x}/{y}.png"), zoom=5, xname=c("BaseTile","overlaytile"))
}
}
\author{
Chris Reudenbach
}

