% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAmapviewClasses.R, R/mapView.R
\docType{class}
\name{mapview-class}
\alias{mapview-class}
\alias{mapView,stars-method}
\title{Class mapview}
\usage{
\S4method{mapView}{stars}(x, map = NULL,
  maxpixels = mapviewGetOption("mapview.maxpixels"),
  col.regions = mapviewGetOption("raster.palette")(256), at = NULL,
  na.color = mapviewGetOption("na.color"), use.layer.names = FALSE,
  values = NULL, map.types = mapviewGetOption("basemaps"),
  alpha.regions = 0.8, legend = mapviewGetOption("legend"),
  legend.opacity = 1, trim = TRUE, verbose = mapviewGetOption("verbose"),
  layer.name = NULL, homebutton = TRUE, native.crs = FALSE,
  method = c("bilinear", "ngb"), ...)
}
\description{
Class mapview
}
\section{Methods (by generic)}{
\itemize{
\item \code{mapView}: \code{stars}
}}

\section{Slots}{

\describe{
\item{\code{object}}{the spatial object}

\item{\code{map}}{the leaflet map object}
}}

