% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-defunct.R
\name{mapview-defunct}
\alias{mapview-defunct}
\title{Defunct functions in mapview}
\description{
These functions have been removed from package \code{mapview}.
See below for information on which package they have been moved to.
}
\details{
\itemize{
 \item \code{cubeview}: This function is defunct, and has been
 migrated to package 'cubeview'.
 \item \code{cubeView}: This function is defunct, and has been
 migrated to package 'cubeview'.
 \item \code{cubeViewOutput}: This function is defunct, and has been
 migrated to package 'cubeview'.
 \item \code{renderCubeView}: This function is defunct, and has been
 migrated to package 'cubeview'.
 \item \code{slideview}: This function is defunct, and has been
 migrated to package 'slideview'.
 \item \code{slideView}: This function is defunct, and has been
 migrated to package 'slideview'.
 \item \code{slideViewOutput}: This function is defunct, and has been
 migrated to package 'slideview'.
 \item \code{renderslideView}: This function is defunct, and has been
 migrated to package 'slideview'.
 \item \code{latticeView}: This function is defunct, and has been
 migrated to package 'leafsync'.
 \item \code{sync}: This function is defunct, and has been
 migrated to package 'leafsync'.
 \item \code{plainview}: This function is defunct, and has been
 migrated to package 'plainview'.
 \item \code{plainView}: This function is defunct, and has been
 migrated to package 'plainview'.
 \item \code{popupTable}: This function is defunct, and has been
 migrated to package 'leafpop'.
 \item \code{popupImage}: This function is defunct, and has been
 migrated to package 'leafpop'.
 \item \code{popupGraph}: This function is defunct, and has been
 migrated to package 'leafpop'.
 \item \code{addFeatures}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{garnishMap}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addHomeButton}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{removeHomeButton}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addImageQuery}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addLogo}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addMouseCoordinates}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{removeMouseCoordinates}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addStaticLabels}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addExtent}: This function is defunct, and has been
 migrated to package 'leafem'.
 \item \code{addStarsImage}: This function is defunct, and has been
 migrated to package 'leafem'.
}
}
