% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{mapviewColors}
\alias{mapviewColors}
\alias{mapviewPalette}
\alias{mapViewPalette}
\title{mapview version of leaflet::color* functions}
\usage{
mapviewColors(x, zcol = NULL,
  colors = mapviewGetOption("vector.palette"), at = NULL,
  na.color = mapviewGetOption("na.color"), ...)

mapviewPalette(name = "mapviewVectorColors")

mapViewPalette(name)
}
\arguments{
\item{x}{Spatial* or Raster* object}

\item{zcol}{the column to be colored}

\item{colors}{color vector to be used for coloring the levels
specified in at}

\item{at}{numeric vector giving the breakpoints for the colors}

\item{na.color}{the color for NA values.}

\item{...}{additional arguments passed on to \code{\link{level.colors}}}

\item{name}{Name of the color palette to be used. One of
"mapviewVectorColors" (default), "mapviewRasterColors",
"mapviewSpectralColors" or "mapviewTopoColors".}
}
\description{
mapview version of leaflet::color* functions

Color palettes for mapview
}
\seealso{
\code{\link{level.colors}}

\code{\link{colorRampPalette}}
}
\author{
Tim Appelhans
}
