% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{mapviewOutput}
\alias{mapviewOutput}
\title{Create a mapview UI element for use with shiny}
\usage{
mapviewOutput(outputId, width = "100\%", height = 400)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{the width and height of the map
(see \code{\link{shinyWidgetOutput}})}
}
\description{
Create a mapview UI element for use with shiny
}

