% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapviewTheme.R
\name{mapviewTheme}
\alias{mapviewTheme}
\title{A trellis theme to mimick the looks of mapview. This is a modified
version of latticeExtra::theEconomist.theme. (c) Felix Andrews}
\usage{
mapviewTheme(win.fontfamily = NULL, with.bg = FALSE, box = "black", ...)
}
\arguments{
\item{win.fontfamily}{on Windows systems, sets the font by name.}

\item{with.bg}{if TRUE, uses a light blue background and a few other
corresponding changes; otherwise white.}

\item{box}{color for panel boxes, strip outlines, and axis ticks.}

\item{...}{further arguments passed to simpleTheme and used to modify the theme.}
}
\description{
A trellis theme to mimick the looks of mapview. This is a modified
version of latticeExtra::theEconomist.theme. (c) Felix Andrews
}
\examples{
\dontrun{
library(rasterVis)

levelplot(poppendorf[[10]], par.settings = mapviewTheme())
}
}
\seealso{
\code{\link{theEconomist.theme}}
}

