% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moControls.R
\name{moControls}
\alias{createPopupRow}
\alias{createPopupRowAlt}
\alias{createPopupTable}
\alias{moControls}
\title{convenience functions for whisker mustache plugin}
\usage{
createPopupTable(x)

createPopupRow(col.name, value)

createPopupRowAlt(col.name, value)
}
\arguments{
\item{x}{a spatial object}

\item{col.name}{the name of the column of the attribute table}

\item{value}{the corresponding value from the attribute table}
}
\description{
convenience functions for whisker mustache plugin
}
\section{Functions}{
\itemize{
\item \code{createPopupTable}: create popup table for sp objects

\item \code{createPopupRow}: create popup table odd row for sp objects

\item \code{createPopupRowAlt}: create popup table even row for sp objects
}}
\author{
Tim Appelhans
}

