% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{npts}
\alias{npts}
\title{count the number of points/vertices/nodes of sf objects}
\usage{
npts(x)
}
\arguments{
\item{x}{an sf/sfc object}
}
\description{
count the number of points/vertices/nodes of sf objects
}
\examples{
npts(franconia)
npts(sf::st_geometry(franconia[1, ])) # first polygon

npts(breweries) # is the same as
nrow(breweries)

}
