% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plainView.R
\name{plainViewOutput}
\alias{plainViewOutput}
\title{Widget output function for use in Shiny}
\usage{
plainViewOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

