% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfControls.R
\name{pointData.POINT}
\alias{"pointData.POINT"}
\alias{pointData.MULTIPOINT}
\alias{pointData.POINT}
\title{Methods for simple features}
\usage{
\method{pointData}{POINT}(obj)

\method{pointData}{MULTIPOINT}(obj)
}
\arguments{
\item{obj}{a simple features object.}
}
\description{
These functions provide methods for simple features.
}
\section{Methods (by generic)}{
\itemize{
\item \code{pointData}: method for point data
}}
\examples{
\dontrun{
library(sf)

m <- leaflet() \%>\% addTiles()

garnishMap(m, addPolylines, data = st_as_sf(atlStorms2005))
garnishMap(m, addPolygons, data = st_as_sf(gadmCHE))
garnishMap(m, addCircleMarkers, data = st_as_sf(breweries91))

}
}

