% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleControls.R
\name{polygonData.sf}
\alias{"pointData.POINT","pointData.POINT}
\alias{M","pointData.POINT}
\alias{M,pointData.MULTIPOINT}
\alias{Z","pointData.POINT}
\alias{Z,pointData.MULTIPOINT}
\alias{ZM}
\alias{ZM"}
\alias{pointData.MULTIPOINT}
\alias{pointData.MULTIPOINT,pointData.MULTIPOINT}
\alias{pointData.POINT}
\alias{pointData.sf}
\alias{polygonData.sf}
\title{Methods for simple features}
\usage{
\method{polygonData}{sf}(obj)

\method{pointData}{sf}(obj)

\method{pointData}{MULTIPOINT}(obj)

\method{pointData}{POINT}(obj)
}
\arguments{
\item{obj}{a simple features object.}
}
\description{
These functions provide methods for simple features.
}
\section{Methods (by generic)}{
\itemize{
\item \code{pointData}: method for point data

\item \code{pointData}: method for point data

\item \code{pointData}: method for point data
}}
\examples{
\dontrun{
library(sf)

m <- leaflet() \%>\% addTiles()

garnishMap(m, addPolylines, data = as.sf(atlStorms2005))
garnishMap(m, addPolygons, data = as.sf(gadmCHE))
garnishMap(m, addCircleMarkers, data = as.sf(breweries91))

}

}

