% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleControls.R
\name{polygonData.sfc}
\alias{"pointData.sfc_POINT","pointData.POINT}
\alias{M","pointData.POINT}
\alias{M,pointData.MULTIPOINT}
\alias{Z","pointData.POINT}
\alias{Z,pointData.MULTIPOINT}
\alias{ZM}
\alias{ZM"}
\alias{pointData.MULTIPOINT}
\alias{pointData.MULTIPOINT,pointData.MULTIPOINT}
\alias{pointData.sfc}
\alias{pointData.sfc_POINT}
\alias{polygonData.sfc}
\title{Methods for simple features}
\usage{
\method{polygonData}{sfc}(obj)

\method{pointData}{sfc}(obj)

\method{pointData}{MULTIPOINT}(obj)

\method{pointData}{sfc_POINT}(obj)
}
\arguments{
\item{obj}{a simple features object.}
}
\description{
These functions provide methods for simple features.
}
\section{Methods (by generic)}{
\itemize{
\item \code{pointData}: method for point data

\item \code{pointData}: method for point data

\item \code{pointData}: method for point data
}}
\examples{
\dontrun{
library(sf)

m <- leaflet() \%>\% addTiles()

garnishMap(m, addPolylines, data = st_as_sfc(atlStorms2005))
garnishMap(m, addPolygons, data = st_as_sfc(gadmCHE))
garnishMap(m, addCircleMarkers, data = st_as_sfc(breweries91))

}

}

