% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popupGraph.R
\name{popupGraph}
\alias{popupGraph}
\title{Create HTML strings for popup graphs}
\usage{
popupGraph(graphs, ...)
}
\arguments{
\item{graphs}{A \code{list} of figures associated with \code{x}.}

\item{...}{Further arguments passed on to \code{\link{svg}}.}
}
\value{
A \code{list} of HTML strings required to create popup graphs.
}
\description{
Create HTML strings for \code{popup} graphs used as input for
\code{\link{mapview}}.
}
\examples{
library(sp)

data(meuse)
coordinates(meuse) <- ~ x + y
proj4string(meuse) <- CRS("+init=epsg:28992")

## create plots with points colored according to feature id
library(lattice)
p <- xyplot(copper ~ cadmium, data = meuse@data, col = "grey")
p <- mget(rep("p", length(x)))

clr <- rep("grey", length(meuse))
p <- lapply(1:length(p), function(i) {
  clr[i] <- "red"
  update(p[[i]], col = clr)
})

mapview(meuse, popup = popupGraph(p, width = 3, height = 2.5))

}
\seealso{
\code{\link{popupTable}}.
}

