% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,mapview-method}
\alias{print,mapview-method}
\title{Method for printing mapview objects}
\usage{
\S4method{print}{mapview}(x)
}
\arguments{
\item{x}{a mapview object}
}
\description{
Method for printing mapview objects
}

