% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{renderCubeView}
\alias{renderCubeView}
\title{Widget render function for use in Shiny}
\usage{
renderCubeView(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

