% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscView.R
\name{renderfpView}
\alias{renderfpView}
\title{Widget render function for use in Shiny}
\usage{
renderfpView(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

