% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{renderMapview}
\alias{renderMapview}
\title{Render a mapview widget in shiny}
\usage{
renderMapview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression (with quote())?
This is useful if you want to save an expression in a variable}
}
\description{
Render a mapview widget in shiny
}

