% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plainView.R
\name{renderPlainView}
\alias{renderPlainView}
\title{Widget render function for use in Shiny}
\usage{
renderPlainView(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

