% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideView.R
\name{renderslideView}
\alias{renderslideView}
\title{Widget render function for use in Shiny}
\usage{
renderslideView(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

