% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView2.R
\name{renderSlideView2}
\alias{renderSlideView2}
\title{Widget render function for use in Shiny}
\usage{
renderSlideView2(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

