% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView3.R
\name{renderslideView3}
\alias{renderslideView3}
\title{Widget render function for use in Shiny}
\usage{
renderslideView3(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

