% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show,mapview-method}
\alias{show,mapview-method}
\title{Method for printing mapview objects (show)}
\usage{
\S4method{show}{mapview}(object)
}
\arguments{
\item{object}{a mapview object}
}
\description{
Method for printing mapview objects (show)
}

