% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView2.R
\name{slideView2internal}
\alias{slideView2internal}
\title{<Add Title>}
\usage{
slideView2internal(message, width = NULL, height = NULL, filename1 = NULL,
  filename2 = NULL)
}
\description{
<Add Description>
}

