% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView2.R
\name{slideView2Output}
\alias{slideView2Output}
\title{Widget output function for use in Shiny}
\usage{
slideView2Output(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

