% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideView3.R
\docType{methods}
\name{slideView3}
\alias{slideView3}
\alias{slideView3,RasterStackBrick,RasterStackBrick-method}
\title{TODO Compare two images trough interactive swiping overlay}
\arguments{
\item{img1}{a RasterStack/Brick, RasterLayer or path to a .png file}

\item{img2}{a RasterStack/Brick, RasterLayer or path to a .png file}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{colors}{the color palette to be used for visualising RasterLayers}

\item{na.color}{the color to be used for NA pixels}
}
\description{
Two images are overlaid and a slider is provided to interactively
compare the two images in a before-after like fashion. \code{img1} and
\code{img2} can either be two RasterLayers, two RasterBricks/Stacks or
two character strings. In the latter case it is assumed that these
point to .png images on the disk.

This is a modified implementation of http://bl.ocks.org/rfriberg/8327361
}
\examples{
### raster data ###
library(sp)
library(raster)

data(poppendorf)

stck1 <- subset(poppendorf, c(3, 4, 5))
stck2 <- subset(poppendorf, c(2, 3, 4))
slideView3(stck1, stck2)

}
\author{
Tim Appelhans
Stephan Woellauer
}

