% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideView3.R
\name{slideView3internal}
\alias{slideView3internal}
\title{<Add Title>}
\usage{
slideView3internal(message, width = NULL, height = NULL, filename1 = NULL,
  filename2 = NULL)
}
\description{
<Add Description>
}

