% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView3.R
\name{slideView3Output}
\alias{slideView3Output}
\title{Widget output function for use in Shiny}
\usage{
slideView3Output(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

