% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideView.R
\name{slideViewInternal}
\alias{slideViewInternal}
\title{<Add Title>}
\usage{
slideViewInternal(message, width = NULL, height = NULL, filename1 = NULL,
  filename2 = NULL)
}
\description{
<Add Description>
}

