% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot.R
\docType{methods}
\name{spplot}
\alias{spplot}
\alias{spplot,mapview-method}
\title{spplot method for objects of class 'mapview'}
\usage{
\S4method{spplot}{mapview}(obj, ...)
}
\arguments{
\item{x}{an object of class 'mapview'}
}
\description{
this function attempts to produce a static version of the specified
mapview object that looks similar to the interactive version.
}
\examples{
\dontrun{
library(sp)
m1 <- mapview(breweries91, zcol = "zipcode")
spplot(m1)

m2 <- mapview(gadmCHE, zcol = "NAME_1")
spplot(m2)
}

}
\author{
Tim Appelhans
}

