% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{stormtracks}
\alias{stormtracks}
\title{Atlantic Ocean storms 2005}
\format{\code{sf feature collection LINESTRING}}
\description{
Atlantic Ocean storms 2005
}
\details{
This dataset contains storm tracks for selected storms
in the Atlantic Ocean basin for the year 2005
}

