% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncView.R
\name{syncView}
\alias{syncView}
\alias{syncview}
\title{View two or more synchronised mapview or leaflet maps}
\usage{
syncView(...)

syncview(...)
}
\arguments{
\item{...}{any number of mapview or leaflet objects}
}
\description{
This function produces a synced view of two or more maps
}
\section{Functions}{
\itemize{
\item \code{syncview}: alias for ease of typing
}}
\examples{
\dontrun{
library(sp)

data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

m1 <- mapview(meuse, zcol = "soil", burst = TRUE)
m2 <- mapview(meuse, zcol = "lead")

syncView(m1, m2)

## 4 panels
m1 <- mapview(poppendorf[[1]], legend = FALSE)
m2 <- mapview(poppendorf[[2]], legend = FALSE)
m3 <- mapview(poppendorf[[3]], legend = FALSE)
m4 <- mapview(poppendorf[[4]], legend = FALSE)

syncView(m1, m2, m3, m4)
}

}

