% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testprj4.R
\name{testprj4}
\alias{testprj4}
\title{test some widget things}
\arguments{
\item{x}{\code{\link{sp}} SpatialPointDataframe object}

\item{color}{colors as:  (green,red,blue,teal,yellow,random) for the points/polygons/lines}

\item{x}{a \code{\link{raster}}* object}

\item{color}{color (palette) of the points/polygons/lines/pixels}

\item{map.types}{character spcifications for the base maps.
see \url{http://leaflet-extras.github.io/leaflet-providers/preview/}
for available options.}

\item{alpha}{opacity of the raster layer(s) (not implemented yet)}

\item{legend}{should a legend be plotted (not implemented yet)}

\item{legend.opacity}{opacity of the legend (not implemented yet)}

\item{...}{additional arguments passed on to repective functions.
See \code{\link{addRasterImage}}, \code{\link{addCircles}},
\code{\link{addPolygons}}, \code{\link{addPolylines}} for details}
}
\description{
testproj4takes
}
\examples{
\dontrun{
#we need sp and raster ###
 library(sp)
 library(raster)
### Now we go  bigger
### get some SRTM data
  r<-getGeoData(name="SRTM",extent(83.8,27,9,99,34),download=TRUE)
}
}
\author{
Chris Reudenbach
}

