% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{tmeanCHE}
\alias{tmeanCHE}
\title{Mean January temperature for Switzerland}
\format{\code{"RasterLayer-class"}}
\source{
\url{http://worldclim.org}
}
\description{
Mean January temperature for Switzerland
}
\details{


This dataset was modified (resampled) from the example
at \code{\link{getData}}.
}

