% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{trails}
\alias{trails}
\title{Selected hiking trails in Franconia}
\format{
\code{sf feature collection MULTILINESTRING}
}
\source{
Datenquelle: Bayerische Vermessungsverwaltung - \code{www.geodaten.bayern.de}
\url{https://geodaten.bayern.de/opengeodata/OpenDataDetail.html?pn=bvv_wanderwege}
}
\usage{
trails
}
\description{
Selected hiking trails in Franconia
}
\details{
These hiking trails were downloaded on 06/04/2017 from
\url{https://geoportal.bayern.de/bayernatlas}
These data are published by the owner under Creative Commons Namensnennung 4.0 Deed,
see \url{https://creativecommons.org/licenses/by/4.0/deed.de} for details.
}
\keyword{datasets}
