% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visEarthPole.R
\name{visEarthPole}
\alias{visEarthPole}
\title{test some widget things}
\description{
visEarthPole gets the visEarthPole Data for leaflet
}
\examples{
\dontrun{
#we need sp and raster ###
 library(sp)
 library(raster)
### Now we go  bigger
### get some SRTM data
  r<-getGeoData(name<-"SRTM",extent(83.8,27,9,99,34),download=TRUE)
}
}
\author{
Chris Reudenbach
}

