% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLargeFeatures.R
\name{addLargeFeatures}
\alias{addLargeFeatures}
\title{Add moderately large datasets with up to ~100k features to a map.}
\usage{
addLargeFeatures(map, data, color = "#03F", weight = 4, radius = 8,
  opacity = 0.9, fillOpacity = 0.4, canvasOpacity = 0.4,
  group = deparse(substitute(data)), maxFeatures = 5000, ...)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{data}{the data to be added to the map.}

\item{color}{color of the features. This can be a single character value,
a vector of character values or a function that takes argument \code{n} to
create a vector of \code{n} colors.}

\item{weight}{the weight of the lines.}

\item{radius}{the radius of the circleMarkers (ignored for lines/polygons).}

\item{opacity}{the opacity of the stroke paths.}

\item{fillOpacity}{opacity of the fill (for circleMarkers and polygons).}

\item{canvasOpacity}{the opacity of features when rendered on canvas.}

\item{group}{the name of the group the data layer should belong to.
THIS NEEDS TO BE A VALID CHARACTER STRING, I.E. CANNOT BE 'NA' OR 'NULL'!}

\item{maxFeatures}{the maximum number of features to be drawn using svg.
If for a given zoom the number of visible features is greater than maxFeatures
things will be rendered on canvas and won't be queryable. Tweak (reduce) this
threshold in case your map becomes unresponsive when zooming in. This is
mainly an issue for lines and polygons with many vertices, e.g. polygons
derived from a raster.}

\item{...}{currently not used.}
}
\description{
This function allows users to add moderately sized datasets to a leaflet
or mapview map. Things are drawn on a html canvas for performance. Feature
querying is supported but only at higher zoom levels to preserve performance.
}
\examples{
\dontrun{
library(mapview)
library(sp)
library(sf)
library(ggmap)

data(crime)
crime <- crime[complete.cases(crime), ]

## sp version
coordinates(crime) <- ~ lon + lat
proj4string(crime) <- "+init=epsg:4326"

leaflet() \%>\%
addProviderTiles("CartoDB.Positron") \%>\%
  addLargeFeatures(crime, group = "crime") \%>\%
  addMouseCoordinates() \%>\%
  addLayersControl(overlayGroups = "crime", position = "topleft")

## sf version
crime_sf <- st_as_sf(crime, coords = c("lon", "lat"), crs = "+init=epsg:4326")

leaflet() \%>\%
addProviderTiles("CartoDB.Positron") \%>\%
  addLargeFeatures(crime_sf, group = "crime", color = "black") \%>\%
  addMouseCoordinates() \%>\%
  addLayersControl(overlayGroups = c("crime", "crime2"), position = "topleft") \%>\%
  addLargeFeatures(crime_sf, group = "crime2")

### or use mapview which uses this by default for large features
mapview(crime, zcol = "offense", cex = 3)

}


}

