% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLargeFeatures.R
\name{addLargeFeatures}
\alias{addLargeFeatures}
\title{Add moderately large datasets with up to ~100k features to a map.}
\usage{
addLargeFeatures(map, data, color = "#03F", weight = 4, radius = 8,
  opacity = 0.9, fillOpacity = 0.4, canvasOpacity = 0.4, group = NULL,
  maxFeatures = 5000, ...)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{data}{the data to be added to the map.}

\item{color}{color of the features. This can be a single character value,
a vector of character values or a function that takes argument \code{n} to
create a vector of \code{n} colors.}

\item{weight}{the weight of the lines.}

\item{radius}{the radius of the circleMarkers (ignored for lines/polygons).}

\item{opacity}{the opacity of the stroke paths.}

\item{fillOpacity}{opacity of the fill (for circleMarkers and polygons).}

\item{group}{the name of the group the data layer should belong to.}
}
\description{
This function allows users to add moderately sized datasets to a leaflet
or mapview map. Things are drawn on a html canvas for performance. Feature
querying is supported but only at higher zoom levels to preserve performance.
}
\examples{
\dontrun{
library(mapview)
library(ggplot2)

### blow diamonds up a bit
big <- data.frame(diamonds[rep(seq_len(nrow(diamonds)), 2),])
big$cut <- as.character(big$cut)
big$color <- as.character(big$color)
big$clarity <- as.character(big$clarity)

### provide some random positions
big$x <- rnorm(nrow(big), 0, 10)
big$y <- rnorm(nrow(big), 0, 10)
coordinates(big) <- ~x+y
proj4string(big) <- CRS("+init=epsg:4326")

leaflet() \%>\%
addProviderTiles("CartoDB.Positron") \%>\%
  addLargeFeatures(big, group = "big") \%>\%
  addLayersControl(overlayGroups = "big", position = "topleft")
}


}

