% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{addMouseCoordinates}
\alias{addMouseCoordinates}
\alias{removeMouseCoordinates}
\title{Add mouse coordinate information at top of map.}
\usage{
addMouseCoordinates(map, epsg = NULL, proj4string = NULL, native.crs = FALSE)

removeMouseCoordinates(map)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{epsg}{the epsg string to be shown.}

\item{proj4string}{the proj4string to be shown.}

\item{native.crs}{logical. whether to use the native crs in the coordinates box.}
}
\description{
These functions are deprecated.
Please use leafem::\code{\link[leafem]{addMouseCoordinates}}
and leafem::\code{\link[leafem]{removeMouseCoordinates}} instead.
}
\section{Functions}{
\itemize{
\item \code{removeMouseCoordinates}: remove mouse coordinates information from a map
}}

