% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{addMouseCoordinates}
\alias{addMouseCoordinates}
\alias{removeMouseCoordinates}
\title{Add mouse coordinate information at top of map.}
\usage{
addMouseCoordinates(map, style = c("detailed", "basic"), epsg = NULL,
  proj4string = NULL, native.crs = FALSE)

removeMouseCoordinates(map)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{style}{whether to show 'detailed' or 'basic' mouse position info.
See Details for an explanation.}

\item{epsg}{the epsg string to be shown.}

\item{proj4string}{the proj4string to be shown.}

\item{native.crs}{logical. whether to use the native crs in the coordinates box.}
}
\description{
This function adds a box displaying the current cursor location
(latitude, longitude and zoom level) at the top of a rendered
mapview or leaflet map. In case of mapview, this is automatically added.
NOTE: The information will only render once a mouse movement has happened
on the map.

Remove mouse coordinates information at top of map.
}
\details{
If style is set to "detailed", the following information will be displayed:
\itemize{
  \item x: x-position of the mouse cursor in projected coordinates
  \item y: y-position of the mouse cursor in projected coordinates
  \item epsg: the epsg code of the coordinate reference system of the map
  \item proj4: the proj4 definition of the coordinate reference system of the map
  \item lat: latitude position of the mouse cursor
  \item lon: longitude position of the mouse cursor
  \item zoom: the current zoom level
}

If style is set to "basic", only 'lat', 'lon' and 'zoom' are shown.
}
\section{Functions}{
\itemize{
\item \code{removeMouseCoordinates}: remove mouse coordinates information from a map
}}

\examples{
library(leaflet)

leaflet() \%>\% addProviderTiles("OpenStreetMap") # without mouse position info
leaflet() \%>\%
  addProviderTiles("OpenStreetMap") \%>\%
  addMouseCoordinates(style = "basic") # with basic mouse position info
leaflet() \%>\%
  addProviderTiles("OpenStreetMap") \%>\%
  addMouseCoordinates() # with detailed mouse position info


m = mapview(breweries)
removeMouseCoordinates(m)

}
