% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeView}
\alias{cubeView}
\alias{cubeview}
\title{View a RasterStack or RasterBrick as 3-dimensional data cube.}
\usage{
cubeView(
  x,
  at,
  col.regions = mapviewGetOption("raster.palette"),
  na.color = mapviewGetOption("na.color"),
  legend = TRUE
)

cubeview(...)
}
\arguments{
\item{x}{a RasterStack or RasterBrick}

\item{at}{the breakpoints used for the visualisation. See
\code{\link{levelplot}} for details.}

\item{col.regions}{color (palette).See \code{\link{levelplot}} for details.}

\item{na.color}{color for missing values.}

\item{legend}{logical. Whether to plot a legend.}

\item{...}{currently not used.}
}
\description{
View a RasterStack or RasterBrick as 3-dimensional data cube.
}
\section{Functions}{
\itemize{
\item \code{cubeview}: alias for ease of typing
}}

