% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeView}
\alias{cubeView}
\title{View a RasterStack or RasterBrick as 3-dimensional data cube.}
\usage{
cubeView(x, at, col.regions = mapviewGetOption("raster.palette"))
}
\description{
Create a 3D data cube from a RasterStack or RasterBrick. The cube can be
freely rotated so that Hovmoller views of x - z and y - z are possible.
}
\details{
The visible layers are alterable by keys: \cr
x-axis: LEFT / RIGHT arrow key \cr
y-axis: DOWN / UP arrow key \cr
z-axis: PAGE_DOWN / PAGE_UP key \cr

Note: Because of key focus issues key-press-events are not always
recognised within RStudio at Windows. In this case open the view in
a web-browser (RStudio button: "show in new window").

Press and hold left mouse-button to rotate the cube.
Press and hold right mouse-button to move the cube.
Spin mouse-wheel or press and hold middle mouse-button and
move mouse down/up to zoom the cube.
}
\examples{
\dontrun{
cubeView(poppendorf)
}

}
\author{
Stephan Woellauer and Tim Appelhans
}

