% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeView_RasterBrick}
\alias{cubeView_RasterBrick}
\title{View a RasterBrick as 3-dimensional cube.}
\usage{
cubeView_RasterBrick(x, at, col.regions = mapviewGetOption("raster.palette"))
}
\description{
see details of cubeView: \link{cubeViewRaw}
}
\author{
Stephan Woellauer
}

