% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{cubeViewOutput}
\alias{cubeViewOutput}
\title{Widget output function for use in Shiny}
\usage{
cubeViewOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{the width and height of the map
(see \code{\link{shinyWidgetOutput}})}
}
\description{
Widget output function for use in Shiny
}
