% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{elevation}
\alias{elevation}
\title{Digital elevation model of Franconia}
\format{\code{RasterLayer}}
\source{
\url{http://biogeo.ucdavis.edu/data/diva/msk_alt/DEU_msk_alt.zip}
}
\description{
Digital elevation model of Franconia
}
\details{
Elevation data was downloaded on 17/04/2017 from
\url{http://biogeo.ucdavis.edu/data/diva/msk_alt/DEU_msk_alt.zip}
using raster::getData('alt', country = "DEU") and cropped to the extent
of franconia.
}
