% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{franconia}
\alias{franconia}
\title{Administrative district borders of Franconia}
\format{
\code{sf feature collection MULTIPOLYGON}
}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts}
}
\description{
Administrative district borders of Franconia
}
\details{
The NUTS_2013_01M_SH.zip archive was downloaded on 23/03/2017 from
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts}.
\url{https://gist.github.com/tim-salabim/2845fa90813fa25c18cf83f9b88cbde0}
}
