% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwmap.R
\docType{fwmap}
\name{fwmap}
\alias{fwmap}
\title{Generates an htmlwidget for a fast webGl leaflet map usable for real big data}
\usage{
fwmap(extdata = NULL, extjson = NULL, color = "blue", width = NULL,
  height = NULL, zcol = NULL, map = NULL, burst = FALSE, radius = 10,
  map.types = c("OpenStreetMap", "Esri.WorldImagery",
  "Thunderforest.Landscape"), legend = FALSE, legend.opacity = 1,
  verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(data, env = parent.frame())),
  popup = NULL, ...)
}
\arguments{
\item{extjson}{path and filename of an external json file providing coordinats in lon lat as: [[11.5922235,46.5435654],[11.5986949,46.5361959]]}

\item{color}{colors as:  (green,red,blue,teal,yellow,random) for the points/polygons/lines}

\item{width}{a valid CSS width}

\item{height}{a valid CSS width}

\item{zcol}{attribute name(s) or column number(s) in attribute table
of the column(s) to be rendered}

\item{burst}{whether to show all (TRUE) or only one (FALSE) layers}

\item{radius}{attribute name(s) or column number(s) in attribute table
of the column(s) to be used for defining the size of circles}

\item{data}{a \code{\link{sp}} SpatialPointDataframe object (currently only)}
}
\description{
fwmap ist a first prototype to render big vector data on base of a leaflet map. It uses webGL and htmlwidgets.

This is a modified and adapted implementation of https://github.com/robertleeplummerjr/Leaflet.glify
}
\examples{
### raster data ###
library(sp)
library(raster)

 data(meuse)
 coordinates(meuse) <- ~x+y
 proj4string(meuse) <- CRS("+init=epsg:28992")

 fwmap(data=meuse,color = "random")

 ## Now we try to get a bigger amount of online data
library(ggplot2)
library(sp)
library(rgdal)
library(mapview)
library(profvis)
## profvis of 50000 rows
big <- diamonds[rep(seq_len(nrow(diamonds)), 1),]
big$cut <- as.character(big$cut)
big$color <- as.character(big$color)
big$clarity <- as.character(big$clarity)

big$x <- rnorm(nrow(big), 10, 3)
big$y <- rnorm(nrow(big), 30, 3)

coordinates(big) <- ~x+y
proj4string(big) <- CRS("+init=epsg:4326")

# do it with pure leaflet
leaflet(big) \%>\% addTiles() \%>\% addCircleMarkers(popup = big@data$cut)

# do it with fastmap
fwmap(big)

## now get bigger
big <- diamonds[rep(seq_len(nrow(diamonds)), 30),]
big$cut <- as.character(big$cut)
big$color <- as.character(big$color)
big$clarity <- as.character(big$clarity)

big$x <- rnorm(nrow(big), 10, 3)
big$y <- rnorm(nrow(big), 50, 3)

coordinates(big) <- ~x+y
proj4string(big) <- CRS("+init=epsg:4326")

# map it now
fwmap(extdata = big, color = "random")

# BENCHMARK
system.time(fwmap(extdata = big, color = "random"))

profvis(fwmap(extdata = big, color = "random"))

}
\author{
Chris Reudenbach
}

