% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTile.R
\name{makeTile}
\alias{makeTile}
\title{Create a tile layer from a raster}
\description{
makeTile takes a raster object and creates  a leaflet compatible tile structure.
}
\examples{
\dontrun{
#we need sp and raster ###
 library(sp)
 library(raster)
### Now we go  bigger
### get some SRTM data
  r<-getGeoData(name="SRTM",extent(83.8,27,9,99,34),download=TRUE)
  prj3035<-"proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
  epsg3031<-"+proj=stere +lat_0=-90 +lat_ts=-71 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"
  makeTile(x="mosaicSRTM.tif",outPath="/home/creu/proj/makeTile", t_srs= prj3035,scale=c(1000,5000))
 makeTile(x=fn,outPath="/home/creu/proj/makeTile", s_srs= epsg3031,t_srs= epsg3031)
}
}
\author{
Chris Reudenbach
}

