% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTile.R
\name{makeTile}
\alias{makeTile}
\title{Create a leaflet conform raster tile layer from a GDAL/raster input}
\arguments{
\item{x}{raster object or gdal file}

\item{outPath}{tmpDir()}

\item{scale}{NULL}

\item{gray}{TRUE}

\item{rgb}{FALSE}

\item{rgba}{FALSE}

\item{s_srs}{"+proj=longlat +datum=WGS84"}

\item{t_srs}{"+proj=longlat +datum=WGS84"}

\item{epsg}{NULL}

\item{rType}{"near"}

\item{zoom}{NULL}
}
\description{
makeTile takes a raster object or a GDAL raster file  and creates  a leaflet compatible tile structure in a given directory.
}
\details{
soon more
}
\examples{
\dontrun{
 # we need a running gdal with all python bindings  ###
 require(raster)
 require(gdalUtils)
 require("curl")
 ### get some data for the antartica

  curl_download(url="ftp://ftp.uninett.no/pub/quantarctica/Quantarctica2.zip", destfile="~/tmp/quantartic.zip"  ,  quiet = TRUE, mode = "wb")
  unzip("~/tmp/quantartic.zip")
  # define correct projection
  epsg3031<-"+proj=stere +lat_0=-90 +lat_ts=-71 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"
  # choose a dataset
  fn="~/tmp/quantarctic/ETOPO1_DEM.tif"
  makeTile(fn=fn,outPath="~/proj/LeafletTiles/etopo1/", s_srs= epsg3031,t_srs= epsg3031))
}
}
\author{
Chris Reudenbach
}

