% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapControls.R
\name{mapControls}
\alias{getLayerControlEntriesFromMap}
\alias{getLayerNamesFromMap}
\alias{initBaseMaps}
\alias{initMap}
\alias{layerName}
\alias{layers2bHidden}
\alias{mapControls}
\alias{mapViewLayersControl}
\alias{rasterCheckAdjustProjection}
\alias{scaleCoordinates}
\alias{scaleLinesCoordinates}
\alias{scalePolygonsCoordinates}
\alias{spCheckAdjustProjection}
\title{convenience functions for working with spatial objects and leaflet maps}
\usage{
getLayerControlEntriesFromMap(map)

getLayerNamesFromMap(map)

layers2bHidden(map)

rasterCheckAdjustProjection(x, maxpixels)

initBaseMaps(map.types)

initMap(map, map.types, proj4str)

scaleCoordinates(x.coords, y.coords)

scalePolygonsCoordinates(x)

scaleLinesCoordinates(x)

spCheckAdjustProjection(x, verbose = FALSE)

mapViewLayersControl(map, map.types, names)

layerName()
}
\arguments{
\item{map}{a leaflet map}

\item{x}{a Raster* or Spatial* object}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{map.types}{map types to be used as backgraound maps}

\item{proj4str}{\code{\link{proj4string}} of the spatial objects to be viewed}

\item{x.coords}{vector of x coordinates}

\item{y.coords}{vector of y coordinates}

\item{verbose}{whether details should be printed to the console}

\item{names}{names of the layer groups to be added to control button}
}
\description{
convenience functions for working with spatial objects and leaflet maps
}
\section{Functions}{
\itemize{
\item \code{getLayerControlEntriesFromMap}: query leaflet map for position of 'addLayersControl' entry

\item \code{getLayerNamesFromMap}: get layer names of leaflet map

\item \code{layers2bHidden}: identify layers to be hidden from initial map rendering

\item \code{rasterCheckAdjustProjection}: check and potentially adjust projection of Raster* objects

\item \code{initBaseMaps}: initialise mapView base maps

\item \code{initMap}: initialise mapView map

\item \code{scaleCoordinates}: scale coordinates for unprojected spatial objects

\item \code{scalePolygonsCoordinates}: scale unprojected SpatialPolygons* objects

\item \code{scaleLinesCoordinates}: scale unprojected SpatialLines* objects

\item \code{spCheckAdjustProjection}: check and potentially adjust projection of Spatial* objects

\item \code{mapViewLayersControl}: add leaflet control button to map

\item \code{layerName}: create layer name for grouping in map
}}
\author{
Tim Appelhans
}

