% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapshot.R
\name{mapshot}
\alias{mapshot}
\title{Save mapview or leaflet map as image}
\usage{
mapshot(x, url = NULL, file, ...)
}
\arguments{
\item{x}{\code{mapview} or \code{leaflet} object.}

\item{url}{Filepath of the \code{.html} index file related with \code{x}. If
not supplied, a temporarily created index file will be used.}

\item{file}{Output filename. Currently supported data types are \code{.png},
\code{.pdf}, and \code{.jpeg}.}

\item{...}{Further arguments passed on to \code{\link{webshot}}.}
}
\description{
Save a mapview or leaflet map as \code{.png}, \code{.pdf}, or \code{.jpeg}
image.
}
\examples{
\dontrun{
m <- mapview(breweries91)
mapshot(m, file = "~/map.png")
}

}
\seealso{
\code{\link{webshot}}, \code{\link{saveWidget}}.
}

