% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAAmapviewClasses.R
\docType{class}
\name{mapview-class}
\alias{mapview-class}
\title{Class mapview}
\description{
Class mapview
}
\section{Slots}{

\describe{
\item{\code{object}}{the spatial object}

\item{\code{map}}{the leaflet map object}
}}

