% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{package}
\name{mapview-package}
\alias{mapview-package}
\title{Interactive viewing of spatial objects in R}
\description{
Interactive viewing of spatial objects in R
}
\details{
The package provides functionality to view spatial objects interactively.
The intention is to provide interactivity for easy and quick visualization
during spatial data analysis. It is not intended for fine-tuned presentation
quality map production.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/mapview}
  \item \url{https://r-spatial.github.io/mapview/}
  \item Report bugs at \url{https://github.com/r-spatial/mapview/issues}
}

}
\author{
Tim Appelhans, Florian Detsch, Chris Reudenbach,
Stephan Woellauer, Spaska Forteva, Thomas Nauss,
Environmental Informatics Marburg \cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}
