% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{package}
\name{mapview-package}
\alias{mapview-package}
\alias{mapviewpackage}
\title{GIS-like viewing of spatial objects in R}
\description{
GIS-like viewing of spatial objects in R
}
\details{
The package provides functionality to view spatial objects interactively
using leaflet.
}
\author{
Tim Appelhans \cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}

