% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{package}
\name{mapview-package}
\alias{mapview-package}
\title{GIS-like viewing of spatial objects in R}
\description{
GIS-like viewing of spatial objects in R
}
\details{
The package provides functionality to view spatial objects interactively.
The intention is to provide interactivity for easy and quick visualization
during spatial data analysis. It is not intended for fine-tuned presentation
quality map production.
}
\author{
Tim Appelhans, Florian Detsch, Chris Reudenbach,
Stephan Woellauer, Spaska Forteva, Thomas Nauss,
Environmental Informatics Marburg \cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}

