% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapView.R
\docType{methods}
\name{mapView}
\alias{mapView}
\alias{mapView,RasterLayer-method}
\alias{mapView,RasterStackBrick-method}
\alias{mapView,Satellite-method}
\alias{mapView,SpatialLines-method}
\alias{mapView,SpatialLinesDataFrame-method}
\alias{mapView,SpatialPixelsDataFrame-method}
\alias{mapView,SpatialPoints-method}
\alias{mapView,SpatialPointsDataFrame-method}
\alias{mapView,SpatialPolygons-method}
\alias{mapView,SpatialPolygonsDataFrame-method}
\alias{mapView,missing-method}
\title{View spatial objects interactively}
\usage{
\S4method{mapView}{RasterLayer}(x, map = NULL,
  maxpixels = mapviewOptions(console = FALSE)$maxpixels,
  color = mapViewPalette(7), na.color = mapviewOptions(console =
  FALSE)$nacolor, use.layer.names = FALSE, values = NULL,
  map.types = mapviewOptions(console = FALSE)$basemaps, layer.opacity = 0.8,
  legend = TRUE, legend.opacity = 1, trim = TRUE,
  verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), ...)

\S4method{mapView}{RasterStackBrick}(x, map = NULL,
  maxpixels = mapviewOptions(console = FALSE)$maxpixels,
  color = mapViewPalette(7), na.color = mapviewOptions(console =
  FALSE)$nacolor, values = NULL, map.types = mapviewOptions(console =
  FALSE)$basemaps, legend = FALSE, legend.opacity = 1, trim = TRUE,
  verbose = mapviewOptions(console = FALSE)$verbose, ...)

\S4method{mapView}{Satellite}(x, ...)

\S4method{mapView}{SpatialPixelsDataFrame}(x, zcol = NULL, ...)

\S4method{mapView}{SpatialPointsDataFrame}(x, zcol = NULL, map = NULL,
  burst = FALSE, color = mapViewPalette(7),
  na.color = mapviewOptions(console = FALSE)$nacolor, radius = 10,
  map.types = mapviewOptions(console = FALSE)$basemaps, legend = FALSE,
  legend.opacity = 1, verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), popup = NULL,
  ...)

\S4method{mapView}{SpatialPoints}(x, map = NULL,
  na.color = mapviewOptions(console = FALSE)$nacolor,
  map.types = mapviewOptions(console = FALSE)$basemaps,
  verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), ...)

\S4method{mapView}{SpatialPolygonsDataFrame}(x, zcol = NULL, map = NULL,
  burst = FALSE, color = mapViewPalette(7),
  na.color = mapviewOptions(console = FALSE)$nacolor, values = NULL,
  map.types = mapviewOptions(console = FALSE)$basemaps, legend = FALSE,
  legend.opacity = 1, weight = 2, verbose = mapviewOptions(console =
  FALSE)$verbose, layer.name = deparse(substitute(x, env = parent.frame())),
  popup = NULL, ...)

\S4method{mapView}{SpatialPolygons}(x, map = NULL,
  na.color = mapviewOptions(console = FALSE)$nacolor,
  map.types = mapviewOptions(console = FALSE)$basemaps, weight = 2,
  verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), ...)

\S4method{mapView}{SpatialLinesDataFrame}(x, zcol = NULL, map = NULL,
  burst = FALSE, color = mapViewPalette(7),
  na.color = mapviewOptions(console = FALSE)$nacolor, values = NULL,
  map.types = mapviewOptions(console = FALSE)$basemaps, legend = FALSE,
  legend.opacity = 1, verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), popup = NULL,
  ...)

\S4method{mapView}{SpatialLines}(x, map = NULL,
  na.color = mapviewOptions(console = FALSE)$nacolor,
  map.types = mapviewOptions(console = FALSE)$basemaps,
  verbose = mapviewOptions(console = FALSE)$verbose,
  layer.name = deparse(substitute(x, env = parent.frame())), ...)

\S4method{mapView}{missing}(map.types = mapviewOptions(console =
  FALSE)$basemaps, easter.egg = FALSE)
}
\arguments{
\item{x}{a \code{\link{raster}}* object}

\item{map}{an optional existing map to be updated/added to}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{color}{color (palette) of the points/polygons/lines/pixels}

\item{na.color}{color for missing values}

\item{use.layer.names}{should layer names of the Raster* object be used?}

\item{values}{a vector of values for the visualisation of the layers.
Per default these are calculated based on the supplied raster* object.}

\item{map.types}{character spcifications for the base maps.
see \url{http://leaflet-extras.github.io/leaflet-providers/preview/}
for available options.}

\item{layer.opacity}{opacity of the raster layer(s)}

\item{legend}{should a legend be plotted}

\item{legend.opacity}{opacity of the legend}

\item{trim}{should the raster be trimmed in case there are NAs on the egdes}

\item{verbose}{should some details be printed during the process}

\item{layer.name}{the name of the layer to be shown on the map}

\item{...}{additional arguments passed on to repective functions.
See \code{\link{addRasterImage}}, \code{\link{addCircles}},
\code{\link{addPolygons}}, \code{\link{addPolylines}} for details}

\item{zcol}{attribute name(s) or column number(s) in attribute table
of the column(s) to be rendered}

\item{burst}{whether to show all (TRUE) or only one (FALSE) layers}

\item{radius}{attribute name(s) or column number(s) in attribute table
of the column(s) to be used for defining the size of circles}

\item{popup}{a character vector of the HTML content for the popups. See
\code{\link{addControl}} for details.}

\item{weight}{line width (see \code{\link{leaflet}} for details)}

\item{easter.egg}{well, you might find out if you set this to TRUE}
}
\description{
this function produces an interactive GIS-like view of the specified
spatial object(s) on top of the specified base maps.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterStackBrick}: \code{\link{stack}} / \code{\link{brick}}

\item \code{Satellite}: \code{\link{satellite}}

\item \code{SpatialPixelsDataFrame}: \code{\link{SpatialPixelsDataFrame}}

\item \code{SpatialPointsDataFrame}: \code{\link{SpatialPointsDataFrame}}

\item \code{SpatialPoints}: \code{\link{SpatialPoints}}

\item \code{SpatialPolygonsDataFrame}: \code{\link{SpatialPolygonsDataFrame}}

\item \code{SpatialPolygons}: \code{\link{SpatialPolygons}}

\item \code{SpatialLinesDataFrame}: \code{\link{SpatialLinesDataFrame}}

\item \code{SpatialLines}: \code{\link{SpatialLines}}

\item \code{missing}: initiate a map without an object
}}
\examples{
\dontrun{
mapView()

### raster data ###
library(sp)
library(raster)

data(meuse.grid)
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) = TRUE
meuse_rst <- stack(meuse.grid)

# raster stack
m1 <- mapView(meuse_rst)
m1

# factorial RasterLayer
m2 <- mapView(raster::as.factor(meuse_rst[[4]]))
m2

# SpatialPixelsDataFrame
mapView(meuse.grid, zcol = "soil")


### point vector data ###
## SpatialPointsDataFrame ##
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

# all layers of meuse
mapView(meuse, burst = TRUE)

# only one layer, all info in popups
mapView(meuse)

## SpatialPoints ##
meuse_pts <- as(meuse, "SpatialPoints")
mapView(meuse_pts)



### overlay vector on top of raster ###
mapView(meuse.grid, zcol = "ffreq") + meuse

### polygon vector data ###
data("gadmCHE")
m <- mapView(gadmCHE)
m

## points on polygons ##
centres <- data.frame(coordinates(gadmCHE))
names(centres) <- c("x", "y")
coordinates(centres) <- ~ x + y
projection(centres) <- projection(gadmCHE)
m + centres

### lines vector data
data("atlStorms2005")
mapView(atlStorms2005)
mapView(atlStorms2005, burst = TRUE)
}

}
\author{
Tim Appelhans
}

