% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapviewColors.R
\name{mapviewColors}
\alias{mapviewColors}
\title{mapview version of leaflet::colorNumeric}
\usage{
mapviewColors(col.regions, at, na.color)
}
\arguments{
\item{col.regions}{color vector to be used for coloring the levels
specified in at}

\item{at}{numeric vector giving the breakpoints for the colors}

\item{na.color}{the color for NA values. Defaults to "transparent"}
}
\description{
mapview version of leaflet::colorNumeric
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{levelplot}}
}

