% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapviewColors.R
\name{mapviewColors}
\alias{mapviewColors}
\title{mapview version of leaflet::color* functions}
\usage{
mapviewColors(x, zcol = NULL, colors = mapviewGetOption("vector.palette"),
  at = NULL, na.color = mapviewGetOption("na.color"), ...)
}
\arguments{
\item{x}{Spatial* or Raster* object}

\item{zcol}{the column to be colored}

\item{colors}{color vector to be used for coloring the levels
specified in at}

\item{at}{numeric vector giving the breakpoints for the colors}

\item{na.color}{the color for NA values.}

\item{...}{additional arguments passed on to \code{\link{level.colors}}}
}
\description{
mapview version of leaflet::color* functions
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{level.colors}}
}

