% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapviewOptions.R
\name{mapviewOptions}
\alias{mapviewOptions}
\title{Global options for the mapview package}
\usage{
mapviewOptions(platform, basemaps, rastersize, maxpixels, verbose, nacolor,
  layerscontrolpos, default = FALSE, console = TRUE)
}
\arguments{
\item{platform}{character. The platform to be used (currently only "leaflet"
is allowed)}

\item{basemaps}{character. The basemaps to be used for rendering data. See
\url{http://leaflet-extras.github.io/leaflet-providers/preview/} for possible
values}

\item{rastersize}{numeric. see the maxBytes argument in \code{\link{addRasterImage}}}

\item{maxpixels}{numeric. The maximum amount of pixels allowed for Raster*
objects to be rendered. Defaults to 500000. Set this higher if you have a
potent machine or are patient enough to wait a little}

\item{verbose}{logical. Many functions in mapview provide details about their
behaviour. Set this to TRUE if you want to see these printed to the console}

\item{nacolor}{character. The default color to be used for NA values.
This is relevant for Raster* objects}

\item{layerscontrolpos}{character. Where should the layer control be
placed? One of "topleft", "topright", "bottomleft", "bottomright".}

\item{default}{logical. If TRUE all options are set to their default values}

\item{console}{logical. Should the options be printed to the console}
}
\value{
list of the current options (invisibly).
If no arguments are provided the options are printed.
}
\description{
To permanently set any of these options, you can add them to
<your R installation>/etc/Rprofile.site>. For example,
to change the default number of pixels to be visualised for Raster* objects,
add a line like this: options(mapviewMaxPixels = 700000) to that file.
}
\examples{
\dontrun{
mapviewOptions()
mapviewOptions(nacolor = "pink")
mapviewOptions()
}


}
\author{
Tim Appelhans
}
\seealso{
\code{\link{rasterOptions}}, \code{\link{options}}
}

