% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapViewPalette.R
\name{mapviewPalette}
\alias{mapViewPalette}
\alias{mapviewPalette}
\title{Default color palette for mapview}
\usage{
mapviewPalette(n, name = "mapviewSpectralColors")

mapViewPalette(n, name = "mapviewSpectralColors")
}
\arguments{
\item{n}{the number of colours to be created}

\item{name}{name of the color palette to be used. One of
"mapviewSpectralColors" (default) or "mapviewTopoColors"}
}
\description{
Default color palette for mapview
}
\section{Functions}{
\itemize{
\item \code{mapViewPalette}: for backward compatibility
}}
\author{
Tim Appelhans
}
\seealso{
\code{\link{colorRampPalette}}
}

