% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapViewPalette.R
\name{mapViewPalette}
\alias{mapViewPalette}
\title{Default color palette for mapview}
\arguments{
\item{n}{the number of colours to be created}
}
\description{
Default color palette for mapview
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{colorRampPalette}}
}

