% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapViewPalette.R
\name{mapViewPalette}
\alias{mapViewPalette}
\title{mapView colour palette}
\usage{
mapViewPalette(n)
}
\arguments{
\item{n}{the number of colours to be created}
}
\description{
mapView colour palette
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{colorRampPalette}}
}

