% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapViewPalette.R
\name{mapviewPalette}
\alias{mapViewPalette}
\alias{mapviewPalette}
\title{Color palettes for mapview}
\usage{
mapviewPalette(name = "mapviewVectorColors")

mapViewPalette(name)
}
\arguments{
\item{name}{Name of the color palette to be used. One of
"mapviewVectorColors" (default), "mapviewRasterColors",
"mapviewSpectralColors" or "mapviewTopoColors".}
}
\description{
Color palettes for mapview
}
\section{Functions}{
\itemize{
\item \code{mapViewPalette}: for backward compatibility
}}
\author{
Tim Appelhans
}
\seealso{
\code{\link{colorRampPalette}}
}

