% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plus.R
\docType{methods}
\name{+}
\alias{+}
\alias{+,leaflet,ANY-method}
\alias{+,mapview,ANY-method}
\alias{+,mapview,mapview-method}
\title{add a layer to a mapview or leaflet map}
\usage{
\S4method{+}{mapview,ANY}(e1, e2)

\S4method{+}{leaflet,ANY}(e1, e2)
}
\arguments{
\item{e1}{the map to which the layer should be added}

\item{e2}{(spatial) object to be added}
}
\description{
add a layer to a mapview or leaflet map
}
\examples{
### raster data ###
library(sp)
library(raster)

data(meuse.grid)
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) = TRUE
meuse_rst <- stack(meuse.grid)

m1 <- mapView(meuse_rst[[3]])

### point vector data ###
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

m2 <- mapView(meuse)

### add two mapview objects
m1 + m2
'+'(m1, m2)

### add spatial object to mapview object
m1 + meuse
'+'(m1, meuse)
}

