% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfControls.R
\name{polygonData.sf}
\alias{"pointData.POINT"}
\alias{"polygonData.LINESTRING"}
\alias{"polygonData.MULTILINESTRING"}
\alias{"polygonData.MULTIPOLYGON"}
\alias{"polygonData.POLYGON"}
\alias{pointData.MULTIPOINT}
\alias{pointData.POINT}
\alias{pointData.sf}
\alias{polygonData.LINESTRING}
\alias{polygonData.MULTILINESTRING}
\alias{polygonData.MULTIPOLYGON}
\alias{polygonData.POLYGON}
\alias{polygonData.sf}
\title{Methods for simple features}
\usage{
\method{polygonData}{sf}(obj)

\method{pointData}{sf}(obj)

\method{pointData}{MULTIPOINT}(obj)

\method{pointData}{POINT}(obj)

\method{polygonData}{POLYGON}(obj)

\method{polygonData}{MULTIPOLYGON}(obj)

\method{polygonData}{LINESTRING}(obj)

\method{polygonData}{MULTILINESTRING}(obj)
}
\arguments{
\item{obj}{a simple features object.}
}
\description{
These functions provide methods for simple features.
}
\section{Methods (by generic)}{
\itemize{
\item \code{pointData}: method for point data

\item \code{pointData}: method for point data

\item \code{pointData}: method for point data

\item \code{polygonData}: method for polygon data

\item \code{polygonData}: method for polygon data

\item \code{polygonData}: method for line data

\item \code{polygonData}: method for line data
}}
\examples{
\dontrun{
library(sf)

m <- leaflet() \%>\% addTiles()

garnishMap(m, addPolylines, data = st_as_sf(atlStorms2005))
garnishMap(m, addPolygons, data = st_as_sf(gadmCHE))
garnishMap(m, addCircleMarkers, data = st_as_sf(breweries91))

}

}

