% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleControls.R
\name{polygonData.sf}
\alias{pointData.sf}
\alias{polygonData.sf}
\title{Methods for simple features}
\usage{
polygonData.sf(obj)

pointData.sf(obj)
}
\arguments{
\item{obj}{a simple features object.}
}
\description{
These functions provide methods for simple features.
}
\section{Functions}{
\itemize{
\item \code{pointData.sf}: method for point data
}}
\examples{
\dontrun{
library(sf)

m <- leaflet() \%>\% addTiles()

garnishMap(m, addPolylines, data = as.sf(atlStorms2005))
garnishMap(m, addPolygons, data = as.sf(gadmCHE))
garnishMap(m, addCircleMarkers, data = as.sf(breweries91))

}

}

