% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{poppendorf}
\alias{poppendorf}
\title{Landsat 8 detail of Franconian Switzerland centered on Poppendorf}
\format{\code{"RasterBrick-class"} with 10 bands (layers 8 and 12 have been dropped).}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\description{
Landsat 8 detail of Franconian Switzerland centered on Poppendorf
}
\details{
Use of this data requires your agreement to the USGS regulations on
using Landsat data.
}

