% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popupGraph.R
\name{popupGraph}
\alias{popupGraph}
\title{Create HTML strings for popup graphs}
\usage{
popupGraph(graphs, type = c("png", "svg"), ...)
}
\arguments{
\item{graphs}{A \code{list} of figures associated with \code{x}.}

\item{type}{Output filetype, one of "png" (default) and "svg".}

\item{...}{Further arguments passed on to \code{\link{png}} or
\code{\link{svg}}.}
}
\value{
A \code{list} of HTML strings required to create popup graphs.
}
\description{
Create HTML strings for \code{popup} graphs used as input for
\code{\link{mapview}}.
}
\examples{
\dontrun{
### example: svg -----

library(sp)

data(meuse)
coordinates(meuse) <- ~ x + y
proj4string(meuse) <- CRS("+init=epsg:28992")

## create plots with points colored according to feature id
library(lattice)
p <- xyplot(copper ~ cadmium, data = meuse@data, col = "grey")
p <- mget(rep("p", length(meuse)))

clr <- rep("grey", length(meuse))
p <- lapply(1:length(p), function(i) {
  clr[i] <- "red"
  update(p[[i]], col = clr)
})

mapview(meuse, popup = popupGraph(p, type = "svg", width = 3, height = 2.5))

### example: png -----
pt <- data.frame(x = 174.764474, y = -36.877245)

coordinates(pt) <- ~ x + y
proj4string(pt) <- "+init=epsg:4326"

p2 <- levelplot(t(volcano), col.regions = terrain.colors(100))

mapview(pt, popup = popupGraph(p2, width = 300, height = 400))
}

}
\seealso{
\code{\link{popupTable}}.
}

