% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popupTable.R
\name{popupTable}
\alias{popupTable}
\title{Create mapview popup table}
\usage{
popupTable(x, zcol, use_cpp = TRUE)
}
\arguments{
\item{x}{\code{Spatial*} or \code{Raster*} object.}

\item{zcol}{\code{numeric} or \code{character} vector indicating the columns
included in the output popup table. If missing, all columns are displayed.}

\item{use_cpp}{\code{logical} determining whether or not to enable
\strong{Rcpp} functionality.}
}
\value{
A \code{list} of HTML strings required to create the feature popup table(s).
}
\description{
Create a \code{popup} table used as input for \code{\link{mapview}}. This
optionally allows the user to include only a subset of feature attributes.
}
\examples{
library(sp)

data(meuse)
coordinates(meuse) <- ~ x + y
proj4string(meuse) <- CRS("+init=epsg:28992")

## include columns 1 and 2 only
mapview(meuse, popup = popupTable(meuse, zcol = 1:2))

}

