% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popupTable.R
\name{popupTable}
\alias{popupTable}
\title{Create HTML strings for popup tables}
\usage{
popupTable(x, zcol, use_cpp = TRUE)
}
\arguments{
\item{x}{A \code{Spatial*} object.}

\item{zcol}{\code{numeric} or \code{character} vector indicating the columns
included in the output popup table. If missing, all columns are displayed.}

\item{use_cpp}{\code{logical} determining whether or not to enable
\strong{Rcpp} functionality.}
}
\value{
A \code{list} of HTML strings required to create feature popup table(s).
}
\description{
Create HTML strings for \code{popup} tables used as input for
\code{\link{mapview}}. This optionally allows the user to include only a
subset of feature attributes.
}
\examples{
\dontrun{

## include columns 1 and 2 only
mapview(franconia, popup = popupTable(franconia, zcol = 1:2))
mapview(breweries, zcol = "founded", legend = TRUE,
        popup = popupTable(breweries, zcol = c("founded", "village")))
leaflet() \%>\% addCircleMarkers(data = breweries)
leaflet() \%>\% addCircleMarkers(data = breweries,
                               popup = popupTable(breweries))
}

}
\seealso{
\code{\link{popupGraph}} and \code{\link{popupImage}}.
}

