% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeView.R
\name{renderCubeView}
\alias{renderCubeView}
\title{Widget render function for use in Shiny}
\usage{
renderCubeView(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression (with quote())?
This is useful if you want to save an expression in a variable}
}
\description{
Widget render function for use in Shiny
}
