% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slideView.R
\docType{methods}
\name{slideView}
\alias{slideView}
\alias{slideView,RasterLayer,RasterLayer-method}
\alias{slideView,RasterStackBrick,RasterStackBrick-method}
\alias{slideView,character,character-method}
\title{Compare two images trhough interactive swiping overlay}
\usage{
\S4method{slideView}{RasterLayer,RasterLayer}(img1, img2,
  colors = mapViewPalette(7), na.color = "#00000000", maxpixels = 5e+05)

\S4method{slideView}{character,character}(img1, img2)
}
\arguments{
\item{img1}{a RasterStack/Brick, RasterLayer or path to a .png file}

\item{img2}{a RasterStack/Brick, RasterLayer or path to a .png file}

\item{colors}{the color palette to be used for visualising RasterLayers}

\item{na.color}{the color to be used for NA pixels}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}
}
\description{
Two images are overlaid and a slider is provided to interactively
compare the two images in a before-after like fashion. \code{img1} and
\code{img2} can either be two RasterLayers, two RasterBricks/Stacks or
two character strings. In the latter case it is assumed that these
point to .png images on the disk.

This is a modified implementation of http://bl.ocks.org/rfriberg/8327361
}
\section{Methods (by class)}{
\itemize{
\item \code{img1 = RasterLayer,img2 = RasterLayer}: 

\item \code{img1 = character,img2 = character}: 
}}
\examples{
### raster data ###
library(sp)
library(raster)

data(poppendorf)

stck1 <- subset(poppendorf, c(3, 4, 5))
stck2 <- subset(poppendorf, c(2, 3, 4))
slideView(stck1, stck2)
}
\author{
Tim Appelhans
}

