% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-package.R
\docType{data}
\name{trails}
\alias{trails}
\title{Selected hiking trails in Franconia}
\format{\code{sf feature collection MULTILINESTRING}}
\source{
Datenquelle: Bayerische Vermessungsverwaltung - www.geodaten.bayern.de
\url{http://www.ldbv.bayern.de/produkte/weitere/opendata.html}
}
\description{
Selected hiking trails in Franconia
}
\details{
These hiking trails were downloaded on 06/04/2017 from
\url{https://geoportal.bayern.de/bayernatlas} and processed using 'inst/extdata/create_trails.R'
These data are published by the owner under Creative Commons Namensnennung 3.0 Deutschland, see
\url{https://creativecommons.org/licenses/by/3.0/de/} for details.
}

