% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viewExtent.R
\name{viewExtent}
\alias{addExtent}
\alias{viewExtent}
\alias{viewExtent,addExtent}
\title{view extent/bbox of spatial objects interactively}
\usage{
viewExtent(x, map = NULL, map.types = c("OpenStreetMap",
  "Esri.WorldImagery"), ...)

addExtent(x, map, ...)
}
\arguments{
\item{x}{either a Raster* object or a Spatial* object}

\item{map}{a leaflet map the extent should be added to. If NULL
standard background layers are cretaed}

\item{map.types}{the map types to be used in case map is NULL}

\item{...}{additional arguments passed on to \code{\link{addRectangles}}}
}
\description{
this function produces an interactive GIS-like view of the extent/bbox
of the supplied spatial object
}
\section{Functions}{
\itemize{
\item \code{addExtent}: 
}}
\examples{
### raster data ###
library(sp)
library(raster)

data(meuse.grid)
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) = TRUE

viewExtent(meuse.grid)
}
\author{
Tim Appelhans
}

