% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visEarthPole.R
\name{visEarthPole}
\alias{visEarthPole}
\title{usecase for local leaflet projections}
\usage{
visEarthPole(dateString="2011-10-04", layerList=c(12,10,11),groupList=NULL,scale=scale500,zoom=5)
}
\arguments{
\item{dateString}{a date in the convienient format "2011-10-04". Basically the retrieve of non existing time slots is corrected to the next existing.}

\item{layerList}{default is (12,10,11). You will find 32 layers  to choose. See Details for more info}

\item{groupList}{default = "500" there are two more "250" and "1000" predifined  group list according to the resolution of the data . if you choose "burst" you will get all layers.}

\item{scale}{set scale groups according to the resolution will be removed options are "scale250","scale500" "scale1000".}

\item{zoom}{set zoom level maximum is 5}
}
\description{
visEarthPole is an usecase interface to the Global Imagery
  Browse Services - GIBS  Basically the projection at the South Pole is EPSG
  3031 and somehow a perfect test implementation of proj4leaflet.
  It is up to now VERY basic and just demonstrate the possibilities of using it along with mapview.
}
\details{
Layerlisting for details pleas look at \url{https://wiki.earthdata.nasa.gov/display/GIBS/GIBS+Available+Imagery+Products}\cr
[1]  "AMSR2_Sea_Ice_Concentration_12km" \cr
[2]  "AMSR2_Sea_Ice_Concentration_25km" \cr
[3]  "AMSR2_Sea_Ice_Brightness_Temp_6km_89H" \cr
[4]  "AMSR2_Sea_Ice_Brightness_Temp_6km_89V" \cr
[5]  "AMSRE_Sea_Ice_Concentration_12km" \cr
[6]  "AMSRE_Snow_Depth_Over_Ice" \cr
[7]  "AMSRE_Sea_Ice_Concentration_25km" \cr
[8]  "AMSRE_Sea_Ice_Brightness_Temp_89H" \cr
[9]  "AMSRE_Sea_Ice_Brightness_Temp_89V" \cr
[10]  "BlueMarble_NextGeneration" \cr
[11]  "BlueMarble_ShadedRelief" \cr
[12]  "BlueMarble_ShadedRelief_Bathymetry" \cr
[13]  "Coastlines" \cr
[14]  "Graticule" \cr
[15]  "MODIS_Terra_Snow_Cover" \cr
[16]  "MODIS_Terra_Sea_Ice" \cr
[17]  "MODIS_Terra_Brightness_Temp_Band31_Day" \cr
[18]  "MODIS_Terra_Brightness_Temp_Band31_Night" \cr
[19]  "MODIS_Terra_CorrectedReflectance_TrueColor" \cr
[20]  "MODIS_Terra_CorrectedReflectance_Bands367" \cr
[21]  "MODIS_Terra_CorrectedReflectance_Bands721" \cr
[22]  "MODIS_Aqua_Snow_Cover" \cr
[23]  "MODIS_Aqua_Sea_Ice" \cr
[24]  "MODIS_Aqua_Brightness_Temp_Band31_Day" \cr
[25]  "MODIS_Aqua_Brightness_Temp_Band31_Night" \cr
[26]  "MODIS_Aqua_CorrectedReflectance_TrueColor" \cr
[27]  "MODIS_Aqua_CorrectedReflectance_Bands721" \cr
[28]  "SCAR_Land_Mask" \cr
[29]  "SCAR_Land_Water_Map \cr"
[30]  "VIIRS_SNPP_CorrectedReflectance_TrueColor" \cr
[31]  "VIIRS_SNPP_CorrectedReflectance_BandsM11-I2-I1" \cr
[32]  "VIIRS_SNPP_CorrectedReflectance_BandsM3-I3-M11" \cr
}
\examples{
\dontrun{

 visEarthPole(groupList="1000",dateString="2014-02-04")
}
}
\author{
Chris Reudenbach
}
\references{
\url{https://wiki.earthdata.nasa.gov/display/GIBS/Global+Imagery+Browse+Services+-+GIBS}\cr
\url{https://wiki.earthdata.nasa.gov/display/GIBS/GIBS+Available+Imagery+Products}\cr
\url{http://map1.vis.earthdata.nasa.gov/twms-antarctic/twms.cgi?request=GetTileService}\cr
\url{https://github.com/kartena/Proj4Leaflet}\cr
}

